Split preview and live SFTP tasks up
N.B. This means you must update your .ftppass file with a `live`
and a `preview` key, instead of the single `key1` key that is
currently there, otherwise the task will fail.

Comment out the `sftp:preview` task and uncomment the `sftp:live`,
`autoshot` and `sftp:screenshot` tasks when you're ready to start
publishing to the live server.